refactor: break down task runner functions
